22 : 글 삭제 인증/권한 처리

+++ src/main/java/com/rest1/domain/post/post/controller/ApiV1PostController.java
@@ -53,9 +53,18 @@ public class ApiV1PostController {
     @DeleteMapping("/{id}")
     @Operation(summary = "글 삭제")
     public RsData<Void> deleteItem(
-            @PathVariable Long id
+            @PathVariable Long id,
+            @RequestHeader("Authorization") @NotBlank @Size(min=30,max=50) String apiKey
     ) {
+
+        String authorization = apiKey.replace("Bearer ", ""); // "Bearer " 접두어 제거
+
+        Member actor = memberService.findByApiKey(authorization).orElseThrow(() -> new ServiceException("401-1", "API키가 올바르지 않습니다." ));
         Post post = postService.findById(id).get();
+
+        //게시물 수정 권한 체크
+        if(!actor.equals(post.getWriter())) throw new ServiceException("403-1", "삭제 권한이 없습니다.");
+
         postService.delete(post);
 
         return new RsData<Void>(
